Rename files to follow naming conventions (#54)
* dist file changes

* style: formatted file names to fit angular convention
