+++ src/aboutpage/about-page.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import { FooterComponent } from '../appfooter/leveragefooter.component';
-import { HeaderComponent } from '../appheader/leverageheader.component';
+import { FooterComponent } from '../appfooter/leverage-footer.component';
+import { HeaderComponent } from '../appheader/leverage-header.component';
 
 @Component({
   selector: 'leverage-about-page',
