+++ src/homepage.component.routes.ts
@@ -1,5 +1,5 @@
 import { HomepageComponent } from './homepage/homepage.component';
-import { AboutPageComponent } from './aboutpage/aboutpage.component';
+import { AboutPageComponent } from './aboutpage/about-page.component';
 import { NotFoundComponent } from './notfound/not-found.component';
 
 export const leverageroutes = [
